kernel_elf_parser::arch
Function get_relocate_pairs

pub fn get_relocate_pairs(
    elf: &ElfFile<'_>,
    elf_base_addr: usize,
) -> Vec<RelocatePair>

Read the relocate pairs from the elf file.

Arguments

elf - The elf file

elf_base_addr - The base address of the elf file if the file will be loaded to the memory

Return

It will return a vector of RelocatePair (from super::RelocatePair) which contains the source address and destination address of the relocation.