kernel_elf_parser::arch
Function relocate_pairs

pub fn relocate_pairs(elf: &ElfFile<'_>, base_addr: usize) -> Vec<RelocatePair>

Read relocate pairs from the elf file.

Arguments

elf - The [xmas_elf::ElfFile] data

base_addr - The base address of the elf file if the file will be loaded to the memory

Return

A vector of super::RelocatePair which contains the source and destination address of the relocation.