kernel_elf_parser
Function elf_base_addr

pub fn elf_base_addr(
    elf: &ElfFile<'_>,
    given_base: usize,
) -> Result<usize, String>

Base address of the ELF file loaded into the memory.

When the ELF file is a position-independent executable, the base address will be decided by the kernel.

Otherwise, the base address is determined by the file, and this field given_base will be ignored.

Arguments

elf - The [xmas_elf::ElfFile] data

given_base - The base address of the ELF file given by the kernel

Return

The real base address for ELF file loaded into the memory.