kernel_elf_parser
Function elf_entry

pub fn elf_entry(elf: &ElfFile<'_>, base_addr: usize) -> VirtAddr

Read the entry point of the elf file

Arguments

elf_data - The elf file data

base_addr - The base address of the elf file if the file will be loaded to the memory

Warning

It can’t be used to parse the elf file which need the dynamic linker, but you can do this by calling this function recursively