kernel_elf_parser

Struct ELFParser

pub struct ELFParser<'a> { /* private fields */ }

A wrapper for the ELF file data with some useful methods.

Implementations

impl<'a> ELFParser<'a>

pub fn new( elf: &'a ElfFile<'_>, interp_base: usize, bias: Option<isize>, uspace_base: usize, ) -> Result<Self, &'static str>

Create a new ELFInfo instance.

Arguments

elf - The ELF file data

interp_base - Address of the interpreter if the ELF file is a dynamic executable

bias - Bias for the base address of the PIE executable.

uspace_base - The lowest address of the user space

Note

If the ELF file is a dynamic executable, the interp_base should be the address of the interpreter, and the address of the ELF file will be elf.base_addr() + bias.

pub fn entry(&self) -> usize

The entry point of the ELF file.

pub fn phnum(&self) -> usize

The number of program headers in the ELF file.

pub fn phent(&self) -> usize

The size of the program header table entry in the ELF file.

pub fn phdr(&self) -> usize

The offset of the program header table in the ELF file.

pub fn base(&self) -> usize

The base address of the ELF file loaded into the memory.

pub fn elf(&self) -> &ElfFile<'_>

The ref of the ELF file data.

pub fn auxv_vector(&self, pagesz: usize) -> [AuxvEntry; 17]

Part of auxiliary vectors from the ELF file.

Arguments

pagesz - The page size of the system

Details about auxiliary vectors are described in https://articles.manugarg.com/aboutelfauxiliaryvectors.html

pub fn ph_load(&self) -> Vec<ELFPH>

Read all self::ELFPH with LOAD type of the elf file.